refactor(core): swap PLUGINS_DIR callers to lyra.core.paths, strip commands export (#977)
- lyra.cli_setup and lyra.core.agent.agent now import from lyra.core.paths
- lyra.commands.__init__ emptied (PLUGINS_DIR was its sole export)
