--- src/lyra/commands/__init__.py
@@ -1,6 +1,0 @@
-from importlib.resources import files
-from pathlib import Path
-
-PLUGINS_DIR: Path = Path(str(files(__name__)))
-
-__all__ = ["PLUGINS_DIR"]
